Define `module.exports` to export globals
Traceur relies on Node’s `module.exports` to export variables from the
scope of an anonymous function to the global namespace, but Meteor
doesn’t have that ability. Instead, this commit defines `module` as an
empty object with an `exports` setter that exports variables from the
local scope upward to the global scope. This commit removes the Grasp
dependency, which should make code compile a bit faster too.